refactor(cli): migrate local-inference.js to TypeScript
Convert bin/lib/local-inference.js (228 lines) to src/lib/local-inference.ts
with typed interfaces for ValidationResult and GpuInfo.

Typed RunCaptureFn for DI pattern used by validateLocalProvider,
getOllamaModelOptions, getDefaultOllamaModel, and validateOllamaModel.

616 CLI tests pass. Coverage ratchet passes.

Relates to #924 (shell consolidation).
